+++ lib/pantheios/services/simple_file_log_service.rb
@@ -6,12 +6,13 @@
 #               ::Pantheios::Services::SimpleFileLogService class
 #
 # Created:      17th June 2015
-# Updated:      4th February 2018
+# Updated:      4th June 2020
 #
 # Home:         http://github.com/synesissoftware/Pantheios-Ruby
 #
 # Author:       Matthew Wilson
 #
+# Copyright (c) 2019-2020, Matthew Wilson and Synesis Information Systems
 # Copyright (c) 2015-2018, Matthew Wilson and Synesis Software
 # All rights reserved.
 #
@@ -60,7 +61,7 @@ module Services
 #
 # NOTE: The *LogService* protocol is implemented by a class that provides
 # the instance methods +severity_logged?(severity : Object) : boolean+ and
-# +log(severity : Object, t : Time, prefix : String, msg : String)+
+# +log(severity : Object, t : Time, prefix : String|Array, msg : String)+
 class SimpleFileLogService
 
 	module SimpleFileLogService_Constants